requirements.txt
python-telegram-bot[job-queue, rate-limiter]==20.7
mixpanel~=4.10.0
mixpanel_py_async==0.3.0
